Change page structure so home page is the new game page
